refactor out food.js dom helpers
[Finishes #148716113]
